refactor(tree): component optimization
