+++ packages/components/tree/hooks/useDraggable.ts
@@ -1,6 +1,6 @@
 import { throttle } from 'lodash-es';
-import { reactive } from '../adapt';
-import { TypeTreeItemState } from '../tree-types';
+import { reactive } from '../utils/adapt';
+import { TypeTreeItemState } from '../types';
 
 export interface TypeDragStates {
   isDragOver: boolean;
