+++ packages/components/tree/hooks/useItemEvents.ts
@@ -1,5 +1,5 @@
-import { TypeCheckboxProps, usePrefixClass } from '../adapt';
-import { TypeEventState, TypeTreeItemState } from '../tree-types';
+import { TypeCheckboxProps, usePrefixClass } from '../utils/adapt';
+import { TypeEventState, TypeTreeItemState } from '../types';
 // 这里封装 tree-item 的一般事件
 // 拖动事件，虚拟滚动事件不要安排到这里
 export default function useItemEvents(state: TypeTreeItemState) {
