+++ packages/components/tree/hooks/useTreeEvents.ts
@@ -1,5 +1,5 @@
-import { TreeProps, TypeTreeState, TypeEventState } from '../tree-types';
-import { getMark, emitEvent } from '../util';
+import { TreeProps, TypeTreeState, TypeEventState } from '../types';
+import { getMark, emitEvent } from '../utils';
 import useTreeAction from './useTreeAction';
 
 // tree 组件一般事件处理
