+++ packages/components/tree/hooks/useTreeScroll.ts
@@ -1,6 +1,6 @@
-import { computed, onMounted, TypeRef, useVirtualScroll, TypeScroll, TreeNode } from '../adapt';
+import { computed, onMounted, TypeRef, useVirtualScroll, TypeScroll, TreeNode } from '../utils/adapt';
 import { get } from 'lodash-es';
-import { TypeTreeState, TypeTimer } from '../tree-types';
+import { TypeTreeState, TypeTimer } from '../types';
 import log from '@tdesign/common-js/log/index';
 
 import type { ComponentScrollToElementParams } from '../../common';
