+++ packages/components/tree/index.ts
@@ -1,9 +1,9 @@
 import _TdTree from './tree';
-import { withInstall } from './adapt';
+import { withInstall } from './utils/adapt';
 
 import './style';
 
 export const Tree = withInstall(_TdTree);
 
-export * from './tree-types';
+export * from './types';
 export default Tree;
