+++ packages/components/tree/tree.tsx
@@ -13,7 +13,7 @@ import {
   TypeTreeOptionData,
   useConfig,
   usePrefixClass,
-} from './adapt';
+} from './utils/adapt';
 import useDragHandle from './hooks/useDragHandle';
 import useTreeAction from './hooks/useTreeAction';
 import useTreeNodes from './hooks/useTreeNodes';
@@ -22,8 +22,8 @@ import useTreeState from './hooks/useTreeState';
 import useTreeStore from './hooks/useTreeStore';
 import useTreeStyles from './hooks/useTreeStyles';
 import props from './props';
-import { TreeNodeState, TreeNodeValue, TypeTreeNodeModel } from './tree-types';
-import { getNode } from './util';
+import { TreeNodeState, TreeNodeValue, TypeTreeNodeModel } from './types';
+import { getNode } from './utils';
 
 // 2022.11.02 tabliang 备注
 // 之前尝试实现了嵌套布局，原本预期嵌套布局能够提升大数据量下，全部渲染节点时的性能表现
@@ -37,11 +37,7 @@ export default defineComponent({
     prop: 'value',
     event: 'change',
   },
-
-  props: {
-    ...props,
-  },
-
+  props,
   setup(props, context) {
     const { t, global } = useConfig('tree');
     const classPrefix = usePrefixClass();
