+++ packages/components/tree/utils/adapt.ts
@@ -2,24 +2,24 @@
 // 除此文件之外的其他组件文件，可从 vue2 项目直接复制到 vue3 项目进行维护
 import * as Vue from 'vue';
 import { Ref, SetupContext, ToRefs, VNode, PropType, ComponentPublicInstance, UnwrapNestedRefs } from 'vue';
-import { CheckboxProps } from '../checkbox';
-import { ClassName, TScroll, Styles, TNode, TreeOptionData, TNodeReturnValue } from '../common';
+import { CheckboxProps } from '../../checkbox';
+import { ClassName, TScroll, Styles, TNode, TreeOptionData, TNodeReturnValue } from '../../common';
 import { TypeTreeEventState as TreeEventState } from '@tdesign/common-js/tree/types';
-import { TdTreeProps, TreeInstanceFunctions } from './type';
-import { VirtualScrollConfig } from '../hooks/useVirtualScrollNew';
-import tdWithInstall from '../utils/withInstall';
-import tdUseVModel from '../hooks/useVModel';
-import tdUseDefaultValue from '../hooks/useDefaultValue';
+import { TdTreeProps, TreeInstanceFunctions } from '../type';
+import { VirtualScrollConfig } from '../../hooks/useVirtualScrollNew';
+import tdWithInstall from '../../utils/withInstall';
+import tdUseVModel from '../../hooks/useVModel';
+import tdUseDefaultValue from '../../hooks/useDefaultValue';
 import { TreeStore } from '@tdesign/common-js/tree/tree-store';
 
 export { ref, reactive, computed, watch, onMounted, toRefs, defineComponent, TransitionGroup } from 'vue';
 export { CaretRightSmallIcon as TdCaretRightSmallIcon } from 'tdesign-icons-vue-next';
-export { Checkbox as TCheckBox } from '../checkbox';
-export { Loading as TLoading } from '../loading';
-export { useConfig, usePrefixClass } from '../hooks/useConfig';
-export { useGlobalIcon } from '../hooks/useGlobalIcon';
-export { default as useLazyLoad } from '../hooks/useLazyLoad';
-export { default as useVirtualScroll } from '../hooks/useVirtualScrollNew';
+export { Checkbox as TCheckBox } from '../../checkbox';
+export { Loading as TLoading } from '../../loading';
+export { useConfig, usePrefixClass } from '../../hooks/useConfig';
+export { useGlobalIcon } from '../../hooks/useGlobalIcon';
+export { default as useLazyLoad } from '../../hooks/useLazyLoad';
+export { default as useVirtualScroll } from '../../hooks/useVirtualScrollNew';
 export { TreeNode, privateKey } from '@tdesign/common-js/tree/tree-node';
 export type TypeVModel = ReturnType<typeof tdUseVModel>;
